move login status into middleware
